Basic game interaction using React
Used create-react-app to get a fully working react environment

Followed some of NetNinja's Magic Memory game Youtube tutorials to get started.

App.js
Initialises game by creating an array of 30 tile objects with 3 layers and randomised colors for each layer

Game play handled

Streaks - called Combos in Tiles - are tracked and reported on

Solution uses a Component SingleTile to handle the display of each tile.

I haven't found the data structure easy to work with. I'd been assuming I'd use a class structure with each cell set up as an object of type class, and then having methods to handle the various inspections required for the game. But I can't see how to integrate that with React yet.

So it still needs a lot of refinement.