Fixed issue with layout when mid-layer matched
When a given element on 2 tiles is matched, then the matching element/s are hidden.

Using display:hide works fine with the back-layer and front-layer, but when the mid-layer is hidden it breaks the positioning of the front-layer.

I have solved this by using an opacity of 0% for the mid-layer when handling the matching class.